Don't further truncate summary
By placing the more marker in a page, the position for truncation
has already been decided. No need to override that.

We can also be pretty sure the marker is in a sensible position so
stripping tags isn't necessary any more, retaining links etc.
(We wouldn't have wanted to truncate and leave unbalanced tags)